refactor 1 of App.tsx into seperate components
